Consolidates two go-cmp wrappers into one
