+++ pkg/util/clusterdata/ingress_profiles_task_test.go
@@ -18,7 +18,7 @@ import (
 	"k8s.io/client-go/kubernetes/fake"
 
 	"github.com/Azure/ARO-RP/pkg/api"
-	"github.com/Azure/ARO-RP/test/util/cmp"
+	"github.com/Azure/ARO-RP/pkg/util/cmp"
 )
 
 func TestIngressProfilesEnricherTask(t *testing.T) {
