+++ pkg/util/cmp/cmp.go
@@ -8,6 +8,8 @@ import (
 	"math/big"
 
 	"github.com/google/go-cmp/cmp"
+
+	"github.com/Azure/ARO-RP/pkg/api"
 )
 
 // Diff is a wrapper for github.com/google/go-cmp/cmp.Diff with extra options
@@ -17,6 +19,7 @@ func Diff(x, y interface{}, opts ...cmp.Option) string {
 		// FIXME: Remove x509CertComparer after upgrading to a Go version that includes https://github.com/golang/go/issues/28743
 		cmp.Comparer(x509CertComparer),
 		cmp.Comparer(bigIntComparer),
+		cmp.AllowUnexported(api.MissingFields{}),
 	)
 
 	return cmp.Diff(x, y, newOpts...)
